pyproject.toml

[project]
name = "airflow-wingman"
version = "0.4.1"
description = "Airflow plugin to enable LLMs chat"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Abhishek Bhakat", email = "abhishek.bhakat@hotmail.com"}
]
dependencies = [
    "airflow-mcp-server==0.5.0",
    "apache-airflow>=2.10.0",
    "pydantic-ai>=0.2.4"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Plugins",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.11",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
GitHub = "https://github.com/abhishekbhakat/airflow-wingman"
Issues = "https://github.com/abhishekbhakat/airflow-wingman/issues"

[project.entry-points."airflow.plugins"]
wingman = "airflow_wingman:WingmanPlugin"

[project.optional-dependencies]
dev = [
    "build>=1.2.2",
    "pre-commit>=4.0.1",
    "ruff>=0.9.2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/airflow_wingman"]

[tool.ruff]
line-length = 200
indent-width = 4
fix = true
preview = true

lint.select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "W",  # pycodestyle warnings
    "C90",  # Complexity
    "C",  # flake8-comprehensions
    "ISC",  # flake8-implicit-str-concat
    "T10",  # flake8-debugger
    "A",  # flake8-builtins
    "UP",  # pyupgrade
]

lint.ignore = [
    "C416",  # Unnecessary list comprehension - rewrite as a generator expression
    "C408",  # Unnecessary `dict` call - rewrite as a literal
    "ISC001",  # Single line implicit string concatenation
    "C901"
]

lint.fixable = ["ALL"]
lint.unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint.mccabe]
max-complexity = 12
